Move handler definition into slack.handler
There's currently a hacky circular import between `models` and `slack.api`. This
fixes that, at the cost of moving handler initialization into
`slack.handler`.
